backend/requirements.txt
# Backend dependencies
fastapi
uvicorn
anthropic
pydantic
python-dotenv
google-generativeai
faiss-cpu
numpy
